custom recipe 기능 프론트엔드와 연결

+++ src/main/java/org/project/foodrecipeserver/repository/CustomRecipeRepository.java
@@ -3,5 +3,5 @@ package org.project.foodrecipeserver.repository;
 import org.project.foodrecipeserver.entity.CustomRecipe;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CustomRecipeRepository extends JpaRepository<CustomRecipe, Long> {
+public interface CustomRecipeRepository extends JpaRepository<CustomRecipe, Integer> {
 }
